Feat : Implementation - ParticleManager And Collision Update
파티클 매니저 추가 및, 객체들간의 충돌 구현+ 파티클이펙트 추가

+++ ANGVIK/EntAmmoManager.cpp
@@ -67,7 +67,7 @@ void EntAmmoManager::Fire(POINTFLOAT pos, int dir)
 			angle *= -1;
 		}
 
-		float moveSpeed = (int)RANDOM(50, 150);
+		float moveSpeed = (float)RANDOM(50, 150);
 
 		vecEntAmmos[i]->IsFire(pos, angle, dir, moveSpeed);
 	}
